WIP Integration tests for graphgen
These call the graphgen binary and confirm that it generates the
same pipeline code that is committed in the examples from the graph
that is committed in that example.

Unfortunately, this is not super simple to do in the rust testing
framework, so I'll need to come back to it.